Remove separation between create-pif and create-pifgz
