+++ products/jbrowse-cli/src/commands/create-pif.test.ts
@@ -3,21 +3,19 @@
  */
 
 import path from 'path'
-
+import fs from 'fs'
 import { setup } from '../testUtil'
 
 const base = path.join(__dirname, '..', '..', 'test', 'data')
 const simplePaf = path.join(base, 'volvox_inv_indels.paf')
 
+const exists = (p: string) => fs.existsSync(p)
+
 describe('create-pif', () => {
+  const fn = path.basename(simplePaf, '.paf') + '.pif.gz'
   setup
-    .command(['create-pif', simplePaf])
-    .it('processes volvox paf', async ctx => {
-      let res = ''
-      for (const obj of ctx.stdout) {
-        res += obj[0]
-      }
-      expect(res.split('\n').length).toBe(15)
-      expect(res).toMatchSnapshot()
+    .command(['create-pifgz', simplePaf, '--out', fn])
+    .it('processes volvox paf', () => {
+      expect(exists(fn)).toBeTruthy()
     })
 })
